tests: Fixes compilation warning.
The warning was noticed on g++ 9.3 on Ubuntu 20.04.

Reviewed-on: http://gerrit2.aldebaran.lan/1776
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>

+++ tests/test_qilang_gencodeutility.cpp
@@ -139,7 +139,7 @@ TYPED_TEST(QiLangSafeMemberAsyncActor, safeMemberAsyncConcurrentCalls)
     ASSERT_TRUE(test::finishesWithValue(result));
   }
 
-  EXPECT_EQ(totalCallCount, ptr->log.size());
+  EXPECT_EQ(static_cast<std::size_t>(totalCallCount), ptr->log.size());
   for (auto&& value : ptr->log)
   {
     EXPECT_EQ("asyncFoo()", value);
